+++ src/ffmpeg.rs
@@ -1,7 +1,7 @@
 /// Muxing support using ffmpeg as a subprocess.
 ///
-/// Also see the alternative method of using ffmpeg via the shared library API, implemented in file
-/// "libav.rs".
+/// Also see the alternative method of using ffmpeg via its "libav" shared library API, implemented
+/// in file "libav.rs".
 
 
 use std::fs::File;
@@ -13,12 +13,20 @@ use tempfile::NamedTempFile;
 use anyhow::{Result, Context, anyhow};
 
 
-fn mux_audio_video_ffmpeg(audio_path: &str, video_path: &str, output_path: &Path) -> Result<()> {
+fn mux_audio_video_ffmpeg(
+    audio_path: &str,
+    video_path: &str,
+    output_path: &Path,
+    ffmpeg_location: Option<String>) -> Result<()> {
     let tmpout = NamedTempFile::new()
        .context("creating temporary output file")?;
     let tmppath = tmpout.path().to_str()
-       .context("obtaining name of temporary file")?;
-    let ffmpeg = Command::new("ffmpeg")
+        .context("obtaining name of temporary file")?;
+    let mut ffmpeg_binary = "ffmpeg".to_string();
+    if let Some(loc) = ffmpeg_location {
+        ffmpeg_binary = loc;
+    }
+    let ffmpeg = Command::new(ffmpeg_binary)
         .args(["-hide_banner", "-nostats",
                "-loglevel", "error",  // or "warning", "info"
                "-y",  // overwrite output file if it exists
@@ -77,9 +85,13 @@ fn mux_audio_video_vlc(audio_path: &str, video_path: &str, output_path: &Path) -
 
 
 // First try ffmpeg subprocess, if that fails try vlc subprocess
-pub fn mux_audio_video(audio_path: &str, video_path: &str, output_path: &Path) -> Result<()> {
-    // eprintln!("Muxing audio {}, video {}", audio_path, video_path);
-    if let Err(e) = mux_audio_video_ffmpeg(audio_path, video_path, output_path) {
+pub fn mux_audio_video(
+    audio_path: &str,
+    video_path: &str,
+    output_path: &Path,
+    ffmpeg_location: Option<String>) -> Result<()> {
+    log::trace!("Muxing audio {}, video {}", audio_path, video_path);
+    if let Err(e) = mux_audio_video_ffmpeg(audio_path, video_path, output_path, ffmpeg_location) {
         log::warn!("Muxing with ffmpeg subprocess failed: {}", e);
         log::info!("Retrying mux with vlc subprocess");
         if let Err(e) = mux_audio_video_vlc(audio_path, video_path, output_path) {